mssql_mcp_js/requirements.txt
pymssql>=2.2.8
python-dotenv>=1.0.0
mcp>=1.0.0